Make status_string and progress into functions
